simplify: extract timeAgo and fetchGitHubActivity to shared helpers
