+++ packages/cli/src/commands/publish.ts
@@ -25,7 +25,7 @@ import {
   Severity,
   evaluateSkillDirectory,
 } from "@skillkit/core";
-import { formatCount } from "../helpers.js";
+import { formatCount, timeAgo, fetchGitHubActivity } from "../helpers.js";
 
 function sanitizeSkillName(name: string): string | null {
   if (!name || typeof name !== "string") return null;
@@ -496,7 +496,7 @@ export class PublishSubmitCommand extends Command {
     const qualityGrade = qualityScore !== null ? getQualityGradeFromScore(qualityScore) : null;
     const qualityBadge = qualityScore !== null ? formatQualityBadge(qualityScore) : "N/A";
 
-    const activity = await this.fetchGitHubActivity(repoInfo.owner, repoInfo.repo);
+    const activity = await fetchGitHubActivity(repoInfo.owner, repoInfo.repo);
 
     const skillEntry = {
       id: `${repoInfo.owner}/${repoInfo.repo}/${skillSlug}`,
@@ -517,7 +517,7 @@ export class PublishSubmitCommand extends Command {
     console.log(colors.muted(`  Quality: ${qualityBadge}${qualityScore !== null ? ` (${qualityScore}/100)` : ""}`));
     if (activity) {
       const stars = formatCount(activity.stars);
-      const pushed = activity.pushedAt ? this.timeAgo(activity.pushedAt) : "unknown";
+      const pushed = activity.pushedAt ? timeAgo(activity.pushedAt) : "unknown";
       console.log(colors.muted(`  Stars: ${stars}`));
       console.log(colors.muted(`  Last push: ${pushed}`));
     }
@@ -658,38 +658,6 @@ export class PublishSubmitCommand extends Command {
       .split(/[-_]/)
       .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
       .join(" ");
-  }
-
-  private async fetchGitHubActivity(
-    owner: string,
-    repo: string,
-  ): Promise<{ stars: number; pushedAt: string | null } | null> {
-    try {
-      const response = await fetch(
-        `https://api.github.com/repos/${owner}/${repo}`,
-        { signal: AbortSignal.timeout(5000) },
-      );
-      if (!response.ok) return null;
-      const data = (await response.json()) as Record<string, unknown>;
-      return {
-        stars: typeof data.stargazers_count === "number" ? data.stargazers_count : 0,
-        pushedAt: typeof data.pushed_at === "string" ? data.pushed_at : null,
-      };
-    } catch {
-      return null;
-    }
-  }
-
-  private timeAgo(dateStr: string): string {
-    const now = Date.now();
-    const then = new Date(dateStr).getTime();
-    if (Number.isNaN(then)) return "unknown";
-    const days = Math.floor((now - then) / 86_400_000);
-    if (days === 0) return "today";
-    if (days === 1) return "yesterday";
-    if (days < 30) return `${days}d ago`;
-    if (days < 365) return `${Math.floor(days / 30)}mo ago`;
-    return `${Math.floor(days / 365)}y ago`;
   }
 
   private createIssueBody(
